Move utility component to src folder
