+++ src/components/utils/CustomLink.tsx
@@ -1,11 +1,11 @@
 import NextLink from "next/link";
 import { Link, LinkProps } from "@chakra-ui/react";
-import { ReactChild, ReactNode } from "react";
+import { ReactNode } from "react";
 // import { useRouter } from "next/router";
 
 type Props = {
   href: string;
-  children: ReactChild | ReactNode;
+  children: ReactNode;
   underline?: boolean;
 };
 
@@ -19,7 +19,7 @@ export default function CustomLink({
   // const isActive = router.pathname === "/";
 
   return (
-    <NextLink href={href} passHref>
+    <NextLink href={href} legacyBehavior passHref>
       <Link
         {...props}
         _hover={{
